+++ lib/hcloud/entities/server_protection.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
 module HCloud
-  class Protection < Entity
+  class ServerProtection < Entity
     attribute :delete, :boolean
+    attribute :rebuild, :boolean
 
     alias delete? delete
+    alias rebuild? rebuild
   end
 end