+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add server Metadata resource
 - Add support for rate limits
 - `Action#resources` now returns a list of resources instead of a list of hashes
+- Separate server protection entity from other protection entity (only server protection includes `rebuild`)
 
 ## HCloud v1.4.0 (2023-01-22)